Get preemptive multitasking partially working
This appears to segfault right now. Oops.

The difficulty here is this:
- The timer IRQ will not fire again unless we actually write to
  GICC_EOIR.
- But, with only one function called per IRQ, it's not really clear
  how we want to do that.

I do think it makes some sense to have the actions for the IRQ in
some sense "collected" and then we simply fire all of them. One
other way to do this would be to have e.g. a bitmap of actions
that triggered and then execute all of those.

Preemptive multitasking though might be the most special in a lot
of ways. It's the only one that should require context switching
to a new context that has interrupts off.

Like, even semaphores or other synchronization primitives shouldn't
actually switch off when we call e.g. sem_signal or cv_signal.

(Although it's not yet clear to me how this interacts with spinlocks).