create "INSTALL_DEVICE" macro for enumerating all devices
What I don't like about this is:
a) it is kind of opaque
b) it is pretty fragile.

INSTALL_DEVICE doesn't actually do much, and whether or not a device
is actually installed is not a matter of whether it is literally present
in the thing_dev object, but rather a matter of whether it called
INSTALL_DEVICE rather than just directly setting the object.

It might be better to design an API that is clearer and also harder
to mess up, e.g. somehow require us to actually call INSTALL_DEVICE
in order to use a device.